moves web controllers under legacy folder
