+++ app/Http/Controllers/Legacy/Web/ImportController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rogue\Http\Controllers\Web;
+namespace Rogue\Http\Controllers\Legacy\Web;
 
 use Carbon\Carbon;
 use League\Csv\Reader;
