+++ app/Http/Controllers/Legacy/Web/PostController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rogue\Http\Controllers\Web;
+namespace Rogue\Http\Controllers\Legacy\Web;
 
 use Rogue\Services\Fastly;
 use Rogue\Services\Legacy\Two\PostService;
