+++ app/Http/Controllers/Legacy/Web/SignupsController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rogue\Http\Controllers\Web;
+namespace Rogue\Http\Controllers\Legacy\Web;
 
 use Rogue\Models\Signup;
 use Rogue\Services\Registrar;
