+++ routes/api.php
@@ -10,7 +10,7 @@
  */
 
 // Assets
-$router->get('images/{post}', 'Web\ImagesController@show');
+$router->get('images/{post}', 'Legacy\Web\ImagesController@show');
 
 // Legacy API Routes
 $router->group(['prefix' => 'api/v1', 'middleware' => ['legacy-auth']], function () {
